+++ lib/decanter.rb
@@ -14,11 +14,7 @@ module Decanter
         else
           raise ArgumentError, "cannot lookup decanter for #{klass_or_sym} with class #{klass_or_sym.class}"
         end + 'Decanter'
-      begin
-        decanter_name.constantize
-      rescue StandardError
-        raise NameError, "uninitialized constant #{decanter_name}"
-      end
+      decanter_name.constantize
     end
 
     def decanter_from(klass_or_string)
@@ -27,11 +23,7 @@ module Decanter
         when Class
           klass_or_string
         when String
-          begin
-            klass_or_string.constantize
-          rescue StandardError
-            raise NameError, "uninitialized constant #{klass_or_string}"
-          end
+          klass_or_string.constantize
         else
           raise ArgumentError, "cannot find decanter from #{klass_or_string} with class #{klass_or_string.class}"
         end