+++ spec/decanter/decanter_core_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
+require 'rails_helper'
 
 describe Decanter::Core do
   let(:dummy) { Class.new(Decanter::Base) }
@@ -16,7 +17,7 @@ describe Decanter::Core do
   end
 
   describe '#input' do
-    let(:name) { [:profile] }
+    let(:name) { :profile }
     let(:parser) { :string }
     let(:options) { {} }
 
@@ -26,7 +27,7 @@ describe Decanter::Core do
       expect(dummy.handlers.key?(name)).to be true
     end
 
-    context 'for multiple values' do
+    context 'for multiple name values' do
       let(:name) { %i(first_name last_name) }
 
       it 'adds a handler for the provided name' do
@@ -43,11 +44,7 @@ describe Decanter::Core do
     end
 
     it 'the handler has default key equal to the name' do
-      expect(dummy.handlers[name][:key]).to eq name.first
-    end
-
-    it 'the handler has name = provided name' do
-      expect(dummy.handlers[name][:name]).to eq name
+      expect(dummy.handlers[name][:key]).to eq name
     end
 
     it 'the handler passes through the options' do
@@ -55,7 +52,7 @@ describe Decanter::Core do
     end
 
     it 'the handler has parser of provided parser' do
-      expect(dummy.handlers[name][:parsers]).to eq parser
+      expect(dummy.handlers[name][:parsers]).to eq [parser]
     end
 
     context 'with key specified in options' do
@@ -84,10 +81,6 @@ describe Decanter::Core do
 
     it 'the handler has default key :profile' do
       expect(dummy.handlers[assoc][:key]).to eq assoc
-    end
-
-    it 'the handler has name = assoc' do
-      expect(dummy.handlers[assoc][:name]).to eq assoc
     end
 
     it 'the handler passes through the options' do
@@ -124,10 +117,6 @@ describe Decanter::Core do
 
     it 'the handler has default key :profile' do
       expect(dummy.handlers[assoc][:key]).to eq assoc
-    end
-
-    it 'the handler has name = assoc' do
-      expect(dummy.handlers[assoc][:name]).to eq assoc
     end
 
     it 'the handler has assoc = provided assoc' do
@@ -174,78 +163,62 @@ describe Decanter::Core do
           .and_return(Array.wrap(parser))
       end
 
-      it 'returns the provided key and value' do
-        expect(dummy.parse(:first_name, nil, 'bar', {})).to eq(first_name: 'bar')
+      it 'returns the provided value' do
+        expect(dummy.parse(nil, 'bar', {})).to eq('bar')
       end
 
       it 'does not call Parser.parsers_for' do
-        dummy.parse(:first_name, nil, 'bar', {})
+        dummy.parse(nil, 'bar', {})
         expect(Decanter::Parser).to_not have_received(:parsers_for)
       end
     end
 
-    context 'when a parser is specified but a required value is not present' do
-      it 'raises an argument error specifying the key' do
-        expect { dummy.parse(:first_name, :foo, nil, required: true) }
-          .to raise_error(ArgumentError, 'No value for required argument: first_name')
-      end
-    end
-
     context 'when one parser is specified' do
-      let(:key) { :afloat }
       let(:val) { 8.0 }
 
       it 'returns the a key-value pair with the parsed value' do
-        expect(dummy.parse(key, :float, val.to_s, {})).to eq(key => val)
+        expect(dummy.parse(:float, val.to_s, {})).to eq(val)
       end
     end
 
     context 'when one parser is specified' do
-      let(:key) { :hash_me_hearties }
       let(:val) { 'a:b,c:d' }
 
       it 'returns the a key-value pair with the parsed value' do
-        expect(dummy.parse(key, :key_value_splitter, val, {})).to eq(key => { 'a' => 'b', 'c' => 'd' })
+        expect(dummy.parse(:key_value_splitter, val, {})).to eq('a' => 'b', 'c' => 'd')
       end
     end
 
     context 'when several parsers are specified' do
-      let(:key) { :afloat }
       let(:val) { 8.0 }
 
       it 'returns the a key-value pair with the parsed value' do
-        expect(dummy.parse(key, %i(string float), val, {})).to eq(key => val)
+        expect(dummy.parse(%i(string float), val, {})).to eq(val)
       end
     end
 
     context 'when a parser with a preparser is specified' do
       PctParser = Class.new(Decanter::Parser::ValueParser) do
-        def self.name
-          'PctParser'
-        end
-      end.tap do |parser|
-        parser.pre :float
+        pre :float
 
-        parser.parser do |val, _options|
+        parser do |val, _options|
           val / 100
         end
       end
 
-      let(:key) { :afloat }
       let(:val) { 8.0 }
 
       it 'returns the a key-value pair with the parsed value' do
-        expect(dummy.parse(key, %i(string pct), val, {})).to eq(key => val / 100)
+        expect(dummy.parse(%i(string pct), val, {})).to eq(val / 100)
       end
     end
 
     context 'when a hash parser and other parsers are specified' do
-      let(:key) { :split_it! }
       let(:val) { 'foo:3.45,baz:91' }
 
       it 'returns the a key-value pairs with the parsed values' do
-        expect(dummy.parse(key, %i(key_value_splitter pct), val, {}))
-          .to eq(split_it!: { 'foo' => 0.0345, 'baz' => 0.91 })
+        expect(dummy.parse(%i(key_value_splitter pct), val, {}))
+          .to eq('foo' => 0.0345, 'baz' => 0.91)
       end
     end
   end
@@ -308,7 +281,7 @@ describe Decanter::Core do
         context 'when the unhandled keys are ignored' do
           it 'does not raise an error' do
             dummy.ignore :foo
-            expect { dummy.unhandled_keys(args) }.to_not raise_error(Decanter::Core::UnhandledKeysError)
+            expect { dummy.unhandled_keys(args) }.not_to raise_error
           end
         end
       end
@@ -323,186 +296,28 @@ describe Decanter::Core do
   end
 
   describe '#handle' do
-    let(:args)    { { foo: 'hi', bar: 'bye' } }
-    let(:name)    { %i(foo bar) }
-    let(:values)  { args.values_at(*name) }
-    let(:handler) { { type: :input, name: name } }
-
-    before(:each) { allow(dummy).to receive(:handle_input).and_return(:foobar) }
-
-    context 'for an input' do
-      it 'calls the handle_input with the handler and extracted values' do
-        dummy.handle(handler, args)
-        expect(dummy)
-          .to have_received(:handle_input)
-          .with(handler, values)
-      end
-
-      it 'returns the results form handle_input' do
-        expect(dummy.handle(handler, args)).to eq :foobar
-      end
-    end
-  end
-
-  describe '#handle_input' do
     let(:name)    { :name }
     let(:parser)  { double('parser') }
     let(:options) { double('options') }
-    let(:args)    { { name => 'Hi', foo: 'bar' } }
-    let(:values)  { args[name] }
-    let(:handler) { { key: name, name: name, parsers: parser, options: options } }
+    let(:values)  { 'Hi' }
+    let(:handler) { { key: name, name: name, parsers: parser, options: options, type: :input } }
 
     before(:each) do
       allow(dummy).to receive(:parse)
     end
 
     it 'calls parse with the handler key, handler parser, values and options' do
-      dummy.handle_input(handler, args)
+      dummy.handle(handler, values)
       expect(dummy)
         .to have_received(:parse)
         .with(parser, values, options)
     end
   end
 
-  describe '#handle_has_one' do
-    let(:output)   { { foo: 'bar' } }
-    let(:handler)  { { key: 'key', options: {} } }
-    let(:values)   { { baz: 'foo' } }
-    let(:decanter) { double('decanter') }
-
-    before(:each) do
-      allow(decanter)
-        .to receive(:decant)
-        .and_return(output)
-
-      allow(dummy)
-        .to receive(:decanter_for_handler)
-        .and_return(decanter)
-    end
-
-    it 'calls decanter_for_handler with the handler' do
-      dummy.handle_has_one(handler, values)
-      expect(dummy)
-        .to have_received(:decanter_for_handler)
-        .with(handler)
-    end
-
-    it 'calls decant with the values on the found decanter' do
-      dummy.handle_has_one(handler, values)
-      expect(decanter)
-        .to have_received(:decant)
-        .with(values)
-    end
-
-    it 'returns an array containing the key, and the decanted value' do
-      expect(dummy.handle_has_one(handler, values))
-        .to match ({ handler[:key] => output })
-    end
-  end
-
-  describe '#handle_has_many' do
-    let(:output)   { [{ foo: 'bar' }, { bar: 'foo' }] }
-    let(:handler)  { { key: 'key', options: {} } }
-    let(:values)   { [{ baz: 'foo' }, { faz: 'boo' }] }
-    let(:decanter) { double('decanter') }
-
-    before(:each) do
-      allow(decanter)
-        .to receive(:decant)
-        .and_return(*output)
-
-      allow(dummy)
-        .to receive(:decanter_for_handler)
-        .and_return(decanter)
-    end
-
-    it 'calls decanter_for_handler with the handler' do
-      dummy.handle_has_many(handler, values)
-      expect(dummy)
-        .to have_received(:decanter_for_handler)
-        .with(handler)
-    end
-
-    it 'calls decant with the each of the values on the found decanter' do
-      dummy.handle_has_many(handler, values)
-      expect(decanter)
-        .to have_received(:decant)
-        .with(values[0])
-      expect(decanter)
-        .to have_received(:decant)
-        .with(values[1])
-    end
-
-    it 'returns an array containing the key, and an array of decanted values' do
-      expect(dummy.handle_has_many(handler, values))
-        .to match ({ handler[:key] => output })
-    end
-  end
-
-  describe '#handle_association' do
-    let(:assoc) { :profile }
-    let(:handler) do
-      {
-        assoc: assoc,
-        key:   assoc,
-        name:  assoc,
-        type:  :has_one,
-        options: {}
-      }
-    end
-
-    before(:each) do
-      allow(dummy).to receive(:handle_has_one)
-    end
-
-    context 'when there is a verbatim matching key' do
-      let(:args) { { assoc => 'bar', :baz => 'foo' } }
-
-      it 'calls handler_has_one with the handler and args' do
-        dummy.handle_association(handler, args)
-        expect(dummy)
-          .to have_received(:handle_has_one)
-          .with(handler, args[assoc])
-      end
-    end
-
-    context 'when there is a matching key for _attributes' do
-      let(:args) { { "#{assoc}_attributes".to_sym => 'bar', :baz => 'foo' } }
-
-      it 'calls handler_has_one with the _attributes handler and args' do
-        dummy.handle_association(handler, args)
-        expect(dummy)
-          .to have_received(:handle_has_one)
-          .with(hash_including(name: "#{assoc}_attributes".to_sym), args[:profile_attributes])
-      end
-    end
-
-    context 'when there is no matching key' do
-      let(:args) { { foo: 'bar', baz: 'foo' } }
-
-      it 'does not call handler_has_one' do
-        dummy.handle_association(handler, args)
-        expect(dummy).to_not have_received(:handle_has_one)
-      end
-
-      it 'returns an empty hash' do
-        expect(dummy.handle_association(handler, args)).to eq({})
-      end
-    end
-
-    context 'when there are multiple matching keys' do
-      let(:args) { { "#{assoc}_attributes".to_sym => 'bar', assoc => 'foo' } }
-
-      it 'raises an argument error' do
-        expect { dummy.handle_association(handler, args) }
-          .to raise_error(ArgumentError, "Handler #{handler[:name]} matches multiple keys: [:profile, :profile_attributes].")
-      end
-    end
-  end
-
   describe '#decant' do
+    subject { dummy.decant(args) }
+
     let(:args) { { foo: 'bar', baz: 'foo' } }
-    let(:subject) { dummy.decant(args) }
     let(:is_required) { true }
 
     let(:input_hash) do
@@ -513,7 +328,7 @@ describe Decanter::Core do
         }
       }
     end
-    let(:handler) { [[:title], input_hash] }
+    let(:handler) { [:title, input_hash] }
     let(:handlers) { [handler] }
 
     before(:each) do
@@ -545,57 +360,44 @@ describe Decanter::Core do
       it 'returns the merged result' do
         expect(subject).to eq args.merge(args)
       end
+
+      context 'with ActionController::Parameters' do
+        let(:params) { ActionController::Parameters.new(args) }
+        subject { dummy.decant(params) }
+
+        it 'passes the args to unhandled keys' do
+          subject
+          expect(dummy).to have_received(:unhandled_keys).with(args)
+        end
+
+        it 'passes the args to handled keys' do
+          subject
+          expect(dummy).to have_received(:handled_keys).with(args)
+        end
+
+        it 'returns the merged result' do
+          expect(subject).to eq args.merge(args)
+        end
+      end
     end
 
     context 'without args' do
       let(:args) { nil }
-      let(:inputs_required) { true }
-      before(:each) do
-        allow(dummy).to receive(:any_inputs_required?).and_return(inputs_required)
-      end
+      let(:is_required) { true }
 
       context 'when at least one input is required' do
         it 'should raise an exception' do
-          expect { subject }.to raise_error(ArgumentError)
+          allow(dummy).to receive(:handlers).and_return(handlers)
+          expect { subject }.to raise_error(Decanter::Core::MissingRequiredInputValue)
         end
       end
 
       context 'when no inputs are required' do
-        let(:inputs_required) { false }
+        let(:is_required) { false }
 
         it 'should return an empty hash' do
           expect(subject).to eq({})
         end
-      end
-    end
-  end
-
-  describe 'any_inputs_required?' do
-    let(:is_required) { true }
-    let(:input_hash) do
-      {
-        key: 'foo',
-        options: {
-          required: is_required
-        }
-      }
-    end
-    let(:handler) { [[:title], input_hash] }
-    let(:handlers) { [handler] }
-    before(:each) do
-      allow(dummy).to receive(:handlers).and_return(handlers)
-    end
-
-    context 'when required' do
-      it 'should return true' do
-        expect(dummy.any_inputs_required?).to be true
-      end
-    end
-
-    context 'when not required' do
-      let(:is_required) { false }
-      it 'should return false' do
-        expect(dummy.any_inputs_required?).to be false
       end
     end
   end
@@ -611,7 +413,7 @@ describe Decanter::Core do
         }
       }
     end
-    let(:handler) { [[:title], input_hash] }
+    let(:handler) { [:title, input_hash] }
     let(:handlers) { [handler] }
     before(:each) { allow(dummy).to receive(:handlers).and_return(handlers) }
 
@@ -621,6 +423,7 @@ describe Decanter::Core do
         expect(result).to be true
       end
     end
+
     context 'when required args are not present' do
       let(:args) { { name: 'Bob' } }
       it 'should return false' do