+++ spec/spec_helper.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
-require 'dotenv'
-Dotenv.load
 # Report Coverage to Code Climate
 require 'simplecov'
-SimpleCov.start
+SimpleCov.start do
+  add_filter '/.rv/'
+  add_filter '/spec/'
+end
 
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'decanter'